ESC-823 move JourneyStore and associated Store trait into persistence package
