+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/persistence/ExchangeRateCache.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.eusubsidycompliancefrontend.persistence
 
 import play.api.libs.json._
 import uk.gov.hmrc.eusubsidycompliancefrontend.persistence.ExchangeRateCache.DefaultCacheTtl
-import uk.gov.hmrc.eusubsidycompliancefrontend.persistence.RepositoryHelpers.dataKeyForType
+import uk.gov.hmrc.eusubsidycompliancefrontend.persistence.PersistenceHelpers.dataKeyForType
 import uk.gov.hmrc.mongo.cache.{CacheIdType, MongoCacheRepository}
 import uk.gov.hmrc.mongo.{CurrentTimestampSupport, MongoComponent}
